requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appium-python-client==1.1.0
    # via -r requirements.in
behave==1.2.6
    # via -r requirements.in
click==7.1.2
    # via pip-tools
importlib-metadata==4.0.1
    # via pep517
parse-type==0.5.2
    # via behave
parse==1.19.0
    # via
    #   behave
    #   parse-type
pep517==0.10.0
    # via pip-tools
pip-tools==6.1.0
    # via -r requirements.in
selenium==3.141.0
    # via
    #   -r requirements.in
    #   appium-python-client
six==1.16.0
    # via
    #   behave
    #   parse-type
toml==0.10.2
    # via pep517
typing-extensions==3.10.0.0
    # via importlib-metadata
urllib3==1.26.4
    # via selenium
zipp==3.4.1
    # via
    #   importlib-metadata
    #   pep517

# The following packages are considered to be unsafe in a requirements file:
# pip
